Remove core/mixins.h.
