+++ include/arviss/rv32/cpus.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "arviss/core/mixins.h"
+#include "arviss/common/types.h"
 #include "arviss/rv32/dispatchers.h"
 #include "arviss/rv32/executors.h"
